+++ src/Database.h
@@ -8,10 +8,10 @@ public:
 	Database();
 	Database(int length);
 	vector<double> fuel_coms;
-	vector<int> rpm;
-	vector<int> temp;
-	vector<int> gear;
-	vector<int> speed;
+	vector<double> rpm;
+	vector<double> temp;
+	vector<double> gear;
+	vector<double> speed;
 
 };
